requirements.txt
Pillow>=10.0.0
PySide6==6.8.0.2
PySide6-Fluent-Widgets>=1.10.0